chore(dockerfile): remove GPL-2.0 license on entire Dockerfiles
Originally, the Dockerfiles that built BusyBox (which uses license GPL-2.0) were wholly licensed under it too in the spirit of copyleft. However, containers from the built images only interact with the BusyBox binaries via `exec` for minor operations, so the respective built images would be more appropriately classified as "aggregate works" and the code to build the BusyBox binaries are separated into their own shell scripts that are licensed under GPL-2.0 and the respective Dockerfiles can execute the script without being considered a derivative work.